pyproject.toml
[project]
name = "grape"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "fastapi[standard]>=0.135.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn>=0.42.0",
]
